add tests and rename test script
